Fold operation lookup tests into queryAst spec
